# docs/swagger.yaml
definitions:
  models.IncludedResources:
    properties:
      loaded_resources:
        items: {}
        type: array
      resources:
        items:
          $ref: '#/definitions/models.Resource'
        type: array
    type: object
  models.Resource:
    properties:
      data: {}
      present:
        type: boolean
      require:
        $ref: '#/definitions/models.ResourceRequire'
      title:
        type: string
      type:
        type: string
      when:
        type: string
    type: object
  models.ResourceRequire:
    properties:
      title:
        type: string
      type:
        type: string
    type: object
  models.Role:
    properties:
      included_resources:
        additionalProperties:
          $ref: '#/definitions/models.IncludedResources'
        type: object
      loaded_resources:
        items: {}
        type: array
      name:
        type: string
      resources:
        items:
          $ref: '#/definitions/models.Resource'
        type: array
      templates:
        additionalProperties:
          type: string
        type: object
      variables:
        additionalProperties: {}
        type: object
    type: object
  requests.RetrieveSignedCertificate:
    properties:
      csr:
        type: string
      node_name:
        type: string
    required:
    - csr
    - node_name
    type: object
  requests.SubmitCertificateRequest:
    properties:
      csr:
        type: string
      node_name:
        type: string
    required:
    - csr
    - node_name
    type: object
  responses.ErrorResponse:
    properties:
      details:
        type: string
      error:
        type: string
    type: object
  responses.GetCatalog:
    properties:
      resources:
        items:
          $ref: '#/definitions/models.Resource'
        type: array
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
      tags:
        items:
          type: string
        type: array
      variables:
        additionalProperties: {}
        type: object
    type: object
  responses.GetRootCA:
    properties:
      certificate:
        type: string
    type: object
  responses.MessageResponse:
    properties:
      details:
        type: string
    type: object
  responses.RetrieveSignedCertificate:
    properties:
      certificate:
        type: string
    type: object
info:
  contact: {}
paths:
  /v1/catalogs/catalog:
    get:
      consumes:
      - application/json
      description: Endpoint use by agent to retrieve the catalog that they need to
        execute
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.GetCatalog'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Retrieve catalog to execute
      tags:
      - catalogs
  /v1/certificates/retrieve:
    post:
      consumes:
      - application/json
      description: retrieve a node signed certificate
      parameters:
      - description: Name of the node, along the CSR
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/requests.RetrieveSignedCertificate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.RetrieveSignedCertificate'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Retrieve a certificate once it has been signed
      tags:
      - certificates
  /v1/certificates/root:
    get:
      description: get root ca from server
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.GetRootCA'
      summary: Used to get the root CA from the server
      tags:
      - certificates
  /v1/certificates/submit:
    post:
      consumes:
      - application/json
      description: submit an unsigned certificate
      parameters:
      - description: Name of the node, along the CSR
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/requests.SubmitCertificateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responses.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Used to submit a certificate
      tags:
      - certificates
swagger: "2.0"